rename Python module to entity_ruler
