+++ spacy/pipeline/entity_ruler.py
